Moving functions from routes file to another file and adding docstring
