+++ app/static/form_config.csv
@@ -1,6 +1,6 @@
 projectId,formId,activityStatus,featureType,lastChecked,lastNumberRecordsMills
-2,Tanzania_Mills_Mapping_Census_V_0.05,0,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=9, tm_hour=15, tm_min=17, tm_sec=4, tm_wday=1, tm_yday=313, tm_isdst=0)",14196
-2,Tanzania_Mills_Mapping_Census_V_0.08,1,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=9, tm_hour=15, tm_min=32, tm_sec=32, tm_wday=1, tm_yday=313, tm_isdst=0)",350
+2,Tanzania_Mills_Mapping_Census_V_0.05,0,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=10, tm_hour=9, tm_min=6, tm_sec=54, tm_wday=2, tm_yday=314, tm_isdst=0)",359
+2,Tanzania_Mills_Mapping_Census_V_0.08,1,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=10, tm_hour=9, tm_min=47, tm_sec=50, tm_wday=2, tm_yday=314, tm_isdst=0)",359
 2,Tanzania_Mills_Mapping_Census_V_0.02,0,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=9, tm_hour=15, tm_min=17, tm_sec=15, tm_wday=1, tm_yday=313, tm_isdst=0)",1405
 2,Tanzania_Mills_Mapping_Census_V_0.06,0,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=9, tm_hour=15, tm_min=18, tm_sec=7, tm_wday=1, tm_yday=313, tm_isdst=0)",11691
-2,Tanzania_Mills_Mapping_Census_V_0.07,1,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=9, tm_hour=15, tm_min=18, tm_sec=31, tm_wday=1, tm_yday=313, tm_isdst=0)",5265
+2,Tanzania_Mills_Mapping_Census_V_0.07,1,point,"time.struct_time(tm_year=2021, tm_mon=11, tm_mday=10, tm_hour=9, tm_min=47, tm_sec=51, tm_wday=2, tm_yday=314, tm_isdst=0)",5265
